Implemented: commong methods for updating count and fetching facilityName information
