+++ src/views/Closed.vue
@@ -107,11 +107,9 @@ import { computed } from "vue"
 import { translate } from "@/i18n";
 import Filters from "@/components/Filters.vue"
 import store from "@/store";
-import { getDateWithOrdinalSuffix, showToast } from "@/utils";
-import router from "@/router";
+import { getCycleCountStats, getDateWithOrdinalSuffix, getFacilityName } from "@/utils";
 
 const cycleCounts = computed(() => store.getters["count/getCounts"])
-const facilities = computed(() => store.getters["user/getFacilities"])
 const cycleCountStats = computed(() => (id: string) => store.getters["count/getCycleCountStats"](id))
 
 onIonViewWillEnter(async () => {
@@ -124,15 +122,6 @@ onIonViewWillLeave(async () => {
   await store.dispatch("count/clearCycleCountList")
   await store.dispatch("count/clearQuery")
 })
-
-function getFacilityName(id: string) {
-  return facilities.value.find((facility: any) => facility.facilityId === id)?.facilityName || id
-}
-
-function getCycleCountStats(id: string) {
-  const stats = cycleCountStats.value(id)
-  return stats ? `${stats.itemCounted}/${stats.totalItems}` : '0/0'
-}
 </script>
 
 <style scoped>
